+++ src/I18n/languages/fr.json
@@ -33,6 +33,7 @@
     "deleteTenantConfirm": "Voulez-vous vraiment supprimer l'organisation '{{tenantName}}'?",
     "restoreTenants": "Restaurer Organisations",
     "restoreTenantsConfirm": "Voulez-vous vraiment restaurer la liste initiale des organisations?",
+    "restoreTenantsSuccess": "La liste des organisations list a été restaurée",
     "noRecordFound": "Pas d'enregistrement trouvé",
     "day": "j",
     "hour": "h",
