+++ src/screens/tenants/CreateTenantDialog.tsx
@@ -4,12 +4,12 @@ import React from 'react';
 import { Alert, TextInput } from 'react-native';
 import Modal from 'react-native-modal';
 
-import computeModalStyleSheet from '../../../ModalStyles';
-import Configuration from '../../../config/Configuration';
-import BaseProps from '../../../types/BaseProps';
-import Tenant, { EndpointCloud, TenantConnection } from '../../../types/Tenant';
-import SecuredStorage from '../../../utils/SecuredStorage';
-import Utils from '../../../utils/Utils';
+import Configuration from '../../config/Configuration';
+import computeModalStyleSheet from '../../ModalStyles';
+import BaseProps from '../../types/BaseProps';
+import { EndpointCloud, TenantConnection } from '../../types/Tenant';
+import SecuredStorage from '../../utils/SecuredStorage';
+import Utils from '../../utils/Utils';
 
 export interface Props extends BaseProps {
   tenants: TenantConnection[];
